Add support for empty dsym associations
Currently when an upload does not create any new dsyms the
association will fail because no checksums are returned.  We now
change the logic so that we just pretend we did an HTTP request
that did not do any assocations.

This way callers do not need to special case.

Fixes #83