feat: Use a connection pool for curl to avoid TLS
We're pretty confident that curl handles stashed into TLS causes all
kinds of mischief on thread shutdown which sometimes might even cause a
segfault.  This instead stashes the handles into an r2d2 connection
pool.

This also removes the manual cleanup now which we added as a workaround
because TLS did not trigger dtors.