feat: Allow setting of Git remote (#637)
Git is decentralised, so it's usual to have several remote
repositories. Origin is usually the default remote repository
but it's not mandatory.

Sentry-cli only allowed to use the remote repository "origin" when
working with commits during a release.

It is now possible to set the remote name in the configuration file
or as environment variable.
In ~/.sentryclirc: defaults.vcs_remote.
In env: SENTRY_VCS_REMOTE.
If not provided, defaults to "origin" so it's backward-compatible.