feat: mv set-manual-commits into set-commits --auto (#772)
We are moving the code in set-manual-commits into set-commits. When you run set-commits --auto if the user does not have a repo based integration, run the set-manual-commits code.
We have agreed to use --local optional flag to represent always using the local git tree to get the commits and to use --initial-depth as the optional flag for the number of commits for the initial release.