fix(chunk_upload): Limit chunk upload waiting to 5 minutes (#896)
Under some conditions, chunk upload may not respond within a given time frame or
even time out completely. This leads sentry-cli to wait forever on a response
from Sentry.

This PR introduces an upper bound of 5 minutes. In chunk upload options, the
Sentry server can override this limit. It is not configurable on the CLI.