fix(dif-upload): Do not supply debug_id for object files
The debug_id can be supplied by the client side according to the
protocol, but there is a protocol bug in that this should have been
gated on the ChunkedUploadCapability::Pdbs feature but is not.

Additionally the server-side does this itself if no debug_id is
supplied and we have more control over fixing the server-side code
than people upgrading sentry-cli, so this is preferable.

Lastly this lucky side-effect is the cause of an incompatibility of
newer sentry-cli versions working with very old sentry servers, kind
of accidentally.

Fixes https://github.com/getsentry/sentry-cli/issues/980 (ish)