ci: dangerbot gently pokes PR authors to include an entry in the changelog (#1034)
This makes it less of a burden on reviewers to check for
and remind authors to add in changelog entries. Uses the
same approach as what's used in other repositories, and
is reflected by the fact that this borrows most of its code 
from symbolicator's implementation with any irrelevant 
bits removed.