Remove `dsyms/associate` API usage
Turns out this API has been deprecated and stubbed out for 5 years now, so it does not make any sense calling it.

I left the PList / Manifest parsing in there for now, as it does validate these files and throws errors. However that can probably also be removed, please advise.