Prefer `--url` argument over empty auth token URL (#1914)
With this change, the CLI will treat an org auth token that specifies an empty URL the same way as an org auth token that is not specifying a URL at all.

We implemented this change in behavior by changing the type of the url in the AuthTokenPayload from an Option<String> to a String. In cases where the url previously would have been set to None, the url will now be set to an empty string. Uses of the AuthTokenPayload's url field have been updated to treat empty strings the way they previously treated None values.

Fixes GH-1913