fix(files): Fail when deleting all files fails with 404 error (#1949)
For some reason, until now, 404 errors received special treatment when running the sentry-cli files delete --all command. Therefore, if for whatever reason, the files could not be deleted due to a 404 error, the command would fail silently. The only indication that anything went wrong would be that the "All files deleted" message would not be printed.

Instead, it would make more sense to print the error message from the Sentry API and fail with a nonzero status, since the CLI has likely failed to accomplish what the user intended for it to accomplish when running the command.

Or, is there any reason why we need to keep the special logic for 404 errors around?