feat: Validate that auth token provided when needed (#1951)
Previously, if users ran a Sentry CLI command that required authentication via an auth token, the CLI would make the API request, which would fail with a 403 error because the user did not provide any authentication. With this change, we locally validate that the auth token is present whenever it is required, and if it is missing, we do not perform the API request.

This change is implemented by creating a new struct called AuthenticatedApi in the api.rs file. The AuthenticatedApi holds a reference to an Api struct, whose config has been verified to have a non-None auth. An authenticated function has been added to the Api struct. This function ensures the Api it is called on has an auth, and if it does, it returns an Ok containing an AuthenticatedApi wrapping the &Api; otherwise, authenticated returns an error.

All high-level Api functions, which call API endpoints requiring authentication (most of the high-level Api functions require authentication), have been moved to AuthenticatedApi. Api now only implements the low-level API functions and the high-level functions that call endpoints which don't require token authentication. All calls to the moved high-level functions have been updated by adding an authenticated()? call before calling methods requiring authentication.

Fixes GH-1905

Co-authored-by: Sebastian Zivota <loewenheim@users.noreply.github.com>