feat(crons): Fully drop auth-token based check in support
We brought this back in GH-1634, since unfortunately at the time we had
many customers still using the legacy check-in endpoints.

We've just disabled the legacy endpoints as of [0] and will be removing
them completely with [1]. This means the token-auth based version of the
monitors run command will no longer work.

[0]: https://github.com/getsentry/sentry-options-automator/pull/851
[1]: https://github.com/getsentry/sentry/pull/66172