ref: Privatize top-level modules and delete dead code (#1992)
Since the CLI is only published as a binary, and not as a library, there is no reason for the top-level modules to be public. So, this PR makes them private.

After making them private, the Rust compiler warned of several parts of the code that were unused. So, this PR removes all the dead code noticed by the compiler. The Rust compiler also warned of some code that was unused in Linux and Windows (but they are still used in the macOS binary), this PR adds #[cfg(target_os = "macos")] directives as needed to ensure this code is only compiled for the macOS platform