feat(api): `files/dsyms/` hits org auth token's region URL (#1997)
When using org auth token authentication, we now direct any API requests to the projects/{}/{}/files/dsyms/ endpoint directly to the region URL encoded in the token. We only do this when the URL is configured to https://sentry.io. For other authentication methods, we default to the old behavior, and log an info message to inform the user that the default URL was used for the upload.