Fix: Improve Xcode error msg when config load fails
When a user sets up a project in xcode, e.g., by using the wizard, the .sentryclirc config is created in the project folder. However, when this file isn't included in the Input Files of the Xcode script for uploading debug files, this causes an error in Xcode which is not very clear. This commit improves that error message.