fix(xcode): Improve Xcode error msg when config load fails (#2028)
When a user sets up a project in xcode, e.g., by using the wizard, the .sentryclirc config is created in the project folder. However, when this file isn't included in the Input Files of the Xcode script for uploading debug files, an error is thrown in Xcode which isn't very clear. This commit improves that error message by specifying what the user needs to do.

Fixes GH-1924