meta: Update version pin README.md (#2123)
The version pin we include in our example for pinning the CLI version is quite old. Let's increase it to the latest version.

Perhaps, in the future, we can automate this change in our release pipeline. For now, though, it is probably enough to make sure we periodically update this, especially after any important fixes.