fix: Only about mismatched URLs when they are different
Previously, we warned about a mismatch between the auth token URL and the manually configured URL even when they were the same. We also sometimes issued the warning when no URL was manually configured (i.e. when the default https://sentry.io/ was used).

Now, we only warn when the two are actually different and we also only warn if a URL was actually manually configured.

Fixes #2137