feat: Log when file not added to source bundle (#2146)
When a file has an invalid (non-UTF-8) encoding, it is skipped from being added to source bundles. Previously, the file was silently skipped. Now, we log when we skip a file.

The logs are emitted at the info level, which is hidden by default (default log level is warn). To see these logs, users will need to set the log level, for example by passing --log-level=info.

Fixes #2135
Fixes #2147