fix(deploys): Honor `--project` in `deploys new` subcommand
Previously, we ignored the `--project` arguments passed to the `sentry-cli deploys new` command. Regardless of whether any `--project` was provided, the deploy was created for all projects associated with the given release.

Now, we respect the `--project` argument. If provided, we only create the deploy for the project(s) specified. If omitted, the deploy is created for all projects associated with the release, as previously.

Fixes #2156