feat(xcode): All uploads in foreground
Perform all uploads from Xcode in the foreground. The new behavior is equivalent to the old `--force-foreground` behavior.

Since uploads are now always executed in the foreground, (soft) deprecate the `--force-foreground` option by hiding it in the command help text. The `--force-foreground` option no longer has any effect, but passing it will continue to be possible at least until the next major release to keep the API backwards compatible.

Background functionality is completely removed from Sentry-CLI with this change.

Closes #2166
Fixes #2148