fix(crons): Run program in `monitors run` even when sending check-in fails
Fix a bug where a failure in sending the crons checkin before running the program in the `monitors run` command (e.g. due to a Sentry outage) would cause the program to be prevented from running. Now, we instead log the error and continue running the program.

Also, ensure that a failure in sending the final checkin (after the program has finished) does not cause the program to exit with a different exit code. `sentry-cli monitors run` should propagate the exit code of the program it runs.

Fixes #2169