fix(crons): Don't panic when passing `--auth-token`
The `--auth-token` argument is hidden from the `sentry-cli monitors run` subcommand since the command uses DSN authentication, however passing the argument is still possible (it is simply ignored).

However, previously providing this argument caused a panic with a not-very-user-friendly message, since the argument was defined here without the auth token parser, and we attempt to read the auth token as an AuthToken.

This PR fixes the bug.

Fixes #2170