ref(api): Only allow dead code where needed (#2213)
We added a blanket `allow(dead_code)` for the entire api module in #2096
because of some new clippy lints which were failing.

Here, we remove the blanket `allow(dead_code)` and only expect dead code
for the specific lines where it occurs. It might not be possible to
remove these because serde might rely on their presence.