test(debug-files): Add test for `assemble` endpoint call
Add an integration test for the `debug-files upload` command, which ensures that given a certain file, the expected initial request to the assemble endpoint is made to the Sentry server. Unlike existing tests, this test primarily does not serve to validate command output, but rather to ensure other behavior (in this case, that a specific HTTP request is made).