test(debug-files): Batch tests together (#2234)
This command only batches together the tests that require no extra
setup. Some tests require per-test setup (specifically most of the
`bundle-jvm` and most of the `upload` tests), so these are kept
separate.

In the future, it likely makes sense to use `assert_cmd` for tests that
require individual setup, and to use `trycmd` for tests which we can
batch together.