test: Run `trycmd` tests with `RUST_BACKTRACE=0` (#2242)
Some of our `trycmd` tests assume that no backtrace is printed in the
test commands output, even when an error is supposed to occur.
`RUST_BACKTRACE=0` appears to be set in CI, but this becomes a problem
running tests locally, where tests sometimes fail when the local
environmnent has `RUST_BACKTRACE=1`.

Note that backtraces are still logged tracing back to the error that
caused the test failure. But, if we are testing that a specific command
which we expect to fail has a certain output, this change ensures that
regardless of the outside environment, we don't get a backtrace _in the
test output_, as this could fail the test.