ref(test): Refactor `EndpointOptions`
Instead of using this struct to store options for the endpoint, which then manually need to be checked in the `mock_endpoint` method, convert the struct into a builder which constructs the mock with the provided configuration.

This will make `EndpointOptions` more flexible, allowing us to expand it with new features, and this also reduces branching.

Keeping `EndpointOptions` around, rather than simply calling `mockito` methods directly, will make it easier to handle any breaking changes to the `mockito` library, such as the ones that will occur when implementing #2223