ref(test): Use `TestManager` in new chunk upload test
Use the `TestManager` in the recently-added `ensure_correct_assemble_call` test. In the future, we would plan to add `assert_cmd` support to `TestManager`, which would make the remaining custom logic in this test redundant, but we will wait until we have a better idea of how this should look like. For now, this change should make #2223 easier to implement