feat(proguard): Introduce experimental chunk uploading feature (#2264)
Introduce an experimental chunk uploading feature for the `sentry-cli
upload-proguard` command. The feature can be activated by setting the
`SENTRY_EXPERIMENTAL_PROGUARD_CHUNK_UPLOAD` environment variable to `1`.
The feature is only activated when users opt in via this environment
variable.

The experimental chunk uploading feature is not backwards compatible. We
attempt the upload regardless of whether the server supports receiving
chunk-uploaded Proguard files. Server-side support will only be
available once https://github.com/getsentry/sentry/pull/81131 is
released. The goal here was to create something that works, so some
optimizations that we use for other chunk uploaded file types (e.g.
first checking whether any chunks are present on the server, and only
uploading the missing ones), are not implemented for Proguard.

Ref #2196