test(chunk-upload): Add test for full chunk upload
Add a test for an entire chunk upload (with only one chunk). The test verifies that the Sentry CLI makes the expected API calls to the assemble endpiont and to the chunk upload endpoint. We check that the request data (including the uploaded file chunk) are exactly the same as what we expect them to be. Our expectation is based on current behavior as of when this change is being made; we assume the current behavior is the correct behavior.

Also, bump MSRV to 1.80 to allow using `LazyLock`, and expect a lint that shows up only after bumping MSRV (I will fix the lint in a separate PR).