test(chunk-upload): Add a test for uploading multiple debug files (#2274)
This test ensures that the correct chunks are sent to the server when
multiple debug files are being uploaded.

Note that our chunk uploading code does not guarantee that the chunks
appear in any particular order within the request. Only the invariant
that all chunks get uploaded (in any arbitrary order) is guaranteed.
Because of this, we need to parse the request body into an unordered set
of all chunks, and compare this against the set of chunks we expect to
receive.

ref #2194