ref(chunks): Rename `Named` trait to `Assemblable`
Rename the `Named` trait to `Assemblable`, in preparation for adding another method to get the Debug ID. We will need to be able to get the DebugID anywhere where we prepare a call to the assemble endpoint, and this is also the only place we need to have a name. So, it makes sense to have a single trait for both.