ref(chunks): Remove `upload-dif` reference from `poll_assemble`
We should not mention `upload-dif` in this error message because (1) `upload-dif` is a soft-deprecated hidden alias to `debug-files upload` and (2) `poll_assemble` has now been generalized and could be used with other commands in the future.

For reason (2), this change makes the error message general rather than referring to the `debug-files upload` command.