ci: Pin Actions runners' OS versions (#2353)
Several of our GitHub Actions runners are running on `ubuntu-latest`,
`windows-latest`, and `macos-latest`, making the actions vulnerable to
breakage when these images are updated. Therefore, we should pin to
specific versions of these images.
    
Pinning to `ubuntu-24.04`, `windows-2022`, and `macos-14` because these
are currently equivalent to `ubuntu-latest`, `windows-latest`, and
`macos-latest`, respectively.