ci(docker): Build each architecture natively
When building the multiarch Docker image, build two separate Docker images for `amd64` and `arm64` natively by using a GitHub Actions runner with the respective architecture (note: `ubuntu-24.04-arm` is currently under public preview, so this could break in the future, but we likely get the best performance by building on the native architecture).

Then, we make the multiarch image by combining the two platform-specific images.

Ref #1338