ci(docker): Release Docker image to GHCR (#2393)
We already build and push our Docker images to GHCR, tagged with commit
hash, during the Release Build action. This PR adds two new actions
which will tag these images with the version number and with the
"latest" once the GitHub release is created. We have two separate
actions for each, since we only want to add the "latest" tag for
releases, not for preleases, but the version tag should be created for
both.

Ref #1503