fix(debug_files): Don't call `xcodebuild`
The code which reads in the `info_plist` (which we never end up using here) internally calls `xcodebuild`. If `xcodebuild` then hangs for whatever reason, then the Sentry CLI will also hang on the upload.

Since we don't need to read the `info_plist` anyway, let's just get rid of it.

Fixes #2440