feat: Remove organization and project info log (#2490)
These appear to have been originally added in #670; however, we have not
adopted these info logs consistently across all of our commands, and we
believe they are likely not very useful.

Since some of these commands are going to be modified to accept multiple
projects, the info log would anyways need to be reworked if we were to
keep it; however, it is simpler just to get rid of the log for now.