Marke all `files` subcommands as deprecated.
The commands are now hidden from help output (making them less likely to be discovered and used), and they are marked with a deprecation notice.

When invoking the commands, a deprecation notice will be printed as well.

Also the previously deprecated `sourcemaps explain` command (which was explained by an in-app sourcemaps debugger)  which already prints a deprecation notice is also being marked as hidden to be less discoverabe.