fix: Use `orig_path` for bundle sources output file name
For dSYMs, `canonical_path` excludes the original dSYM extension, causing any extension before the `.dSYM` to be lost when adding the `.src.zip` extension. This caused the bug observed in #2502.

Instead, we should use the file name given in the original path when generating the file name for the output source bundle.

Fixes #2502