fix: Don't error if invalid value supplied for max retries
Currently, if an invalid value is supplied for the maximum retry count, the CLI will panic, if a command which reads this value is called.

Since we can always fall back to a reasonable default, we should probably warn instead of errroring at all when an invalid value is supplied.

This change is done in preparation for #2209, since that change will add retries to all endpoints, increasing the chance of a user encountering a panic if this behavior is left unfixed.