Setup initial mobile-app command scaffold (#2525)
In preparation for adding mobile application upload support to Sentry,
this creates the basic scaffolding of a new `mobile-app` command with a
single `upload` subcommand. The implementation of this subcommand will
come in subsequent PRs.

Currently, the `mobile-app` command is set to hidden to ensure minimal
exposure if a release is cut.