ref: Disallow using `unwrap` (#2539)
Add a lint which disallows using `unwrap`, requiring an error to be
returned or `expect` to be used, instead.

Since the codebase uses `unwrap` in many places already, which would be
impractical to fix here, this change also suppresses the lint for these
existing violations.

The intention with this change is primarily to enforce better practices
going forward, rather than to try to convert all existing `unwrap` calls
to something else.