build: Fix clippy `unwrap::used` lint in CI (#2543)
Since we run our clippy lints with `--test` in CI, the `unwrap::used`
lint, as currently written, was not triggering (e.g. on the PR #2532,
which contains a violation which should have triggered the lint).

This change should make the `unwrap_used` lint active, also in CI, but
not for test code, as was originally intended.