Add Zip normalization of provided paths before uploading (#2532)
For ease of parsing artifacts on the backend, we'll "normalize" all
provided paths to a basic zip file. This ensures that no matter the
artifact uploaded (files: APK, AAB, directory: XCArchive), we can
process all as a zip file on the backend. This was an optimization done
with Emerge that greatly simplified backend artifact handling.

Skipping tests here to be added at a later time when I leverage a full
integration test to validate the chunked uploads match the expected zip
format.