test(monitors): pass empty options to grep (#2562)
I was trying to run the test suite locally and this particular test was
failing because I override GREP_OPTIONS in my profile.

The test would fail as the actual output would include the shell escape
characters used to display colors in the terminal which are not expected
by the test.

This fixes that.
By the way this is the only place where grep is used so we only need to
update it here.