ref: Exclude `mobile-app` command from release builds (#2582)
`sentry-cli mobile-app` is still under development. As a result, it
makes sense to exclude it from release builds.

This change enables us to still run lints and tests in CI against the
`moblie-app` command, but keeps it out of the release build for now.
Once the command is ready, we can get rid of the feature flag.

Note that the `mobile-app` has not been included in a release yet, so
this change is non-breaking.