feat: allow optionally disabling Swift sandbox
If `cargo` is invoked inside a sandbox on Darwin, the build will fail
when it invokes `swift`. This is because `swift` builds within its own
sandbox by default, but nested sandboxes are not allowed on Darwin.

This is the case, for example, when building `sentry-cli` inside
Homebrew.

For builds that already happen inside a sandbox, allow setting an
environment variable `SWIFT_DISABLE_SANDBOX` which then results in
`swift` being invoked with `--disable-sandbox` to enable the build to
succeed.