feat(js): Add `live: 'rejectOnError'` execution mode to `execute` method (#2605)
The Sentry CLI JS interface's `execute` method takes a `live` parameter
as a second argument:
- if `false`, stdio is returned and the promise returned from `execute`
rejects/resolves based on the CLI binary's exit code.
- if `true`, CLI stdio is piped to the process calling `execute`, but
the returned promise always resolves regardless of the exit code (raised
in #2062)

This feature adds a third value (`'rejectOnError'`) to `live`, which still
pipes stdio but also rejects if the CLI binary's exit code was non-zero.
Arguably, this is what `true` should do but to avoid a breaking change
for now, let's introduce this as a new value instead.