Revert "chore(ci): Use yarn for node tests instead of npm (#2612)" (#2624)
We need to revert installing dependencies with `yarn` because in
contrast to `npm`, `yarn` exits the installation of `optionalDependencies`
if the version does not exist. We're running into this edge case on our
release branch, where we first bump the dependency versions to the next
version to be released and then run tests which try to install this new
version.