build: Delete `Makefile` (#2641)
The `Makefile` was added in
https://github.com/getsentry/sentry-cli/commit/53a4f3b74a33711a8727c25cd425769f28344fcf,
and at that time, it was called from CI. Since then, our CI no longer
calls `make` to build, lint, and test Sentry CLI, so the `Makefile` no
longer appears to serve much of a purpose.

So, let's delete the file to avoid confusing anyone as to its purpose.