feat(logs): Introduce logs command (#2664)
Add a new `logs` subcommand to the Sentry CLI, allowing users to manage
and query logs from their organizations.

This includes a `list` command to fetch and display log entries with
options for filtering and pagination.

Implemented common arguments for logs and integrated them into the
command structure. Also added integration tests for the new
functionality.

Part of https://github.com/getsentry/sentry-cli/issues/2661

Live tailing will be done as a part of separate PR to make this a bit
easier to review, it's already very large change.

---------

Co-authored-by: Simon Hellmayr <simon.hellmayr@sentry.io>