fix(debug-files): Run all processing steps with `--no-upload` (#2693)
According to the docs, `--no-upload` is supposed to "run all steps for
the processing", but currently it exits early and doesn't actually do
any searching for debug files.

This PR defers the exit until _right_ before files are actually
uploaded, which makes the command actually output helpful information
for troubleshooting.

---------

Co-authored-by: Daniel Szoke <daniel.szoke@sentry.io>