feat(launchpad): Add asset catalog files to zip without adding to folder (#2667)
This improves the app upload experience by making it so assets are
written directly to the zip file rather than to the input xcarchive
directory. Previously the assets were written to the input folder and
then all the files in the folder were zipped. This meant the original
location the user passed to the CLI had extra files in it after running
the command, which is confusing.

There was also a bug that assets were not handled for ipa uploads. IPA
files are already zipped, so when we wrote assets to the user's input
directory we only did that for xcarchives (Which is a directory not a
zip file). So a side effect of not writing images to the users input
directory was also to fix this issue for IPAs, and now IPA uploads have
assets included too.