retry on cloudflare timeout (#2695)
**Summary**
Some of the self hosted installations are hosted behind cloudflare. By
default, cloudflare proxy_write_timeout is set at 30s. If this timeout
is breached, Cloudflare responds with error with status code 524.
Currently this is treated as non retriable error by sentry-cli.


https://developers.cloudflare.com/support/troubleshooting/http-status-codes/cloudflare-5xx-errors/error-524/

This PR ensures that sentry-cli retries on this timeout.

Co-authored-by: Manish Rawat <manish@Manishs-MacBook-Air.local>