ci(release): Use `container` in workflow instead of `build-in-docker.sh` (#2727)
This eases maintainability, as the workflow is direcltly defined in
`build.yml`. This will also allow us to add caching of dependencies more
easily, and make the workflow for Linux more closely match the workflow
for macOS and Windows.

We also remove the smoke test, which never triggered anyways, as it was
relying on Travis environment variables.