ci(release): `npm install` in `test_node.yml` on release
`npm ci` will fail here, as the new versions of the optional dependencies are not published yet.

Additionally, add a script to bump the optional dependencies in the package-lock.json after a release is created. Otherwise, `npm ci` will continue to fail after the release, until someone updates the package-lock.json manually.