ci: `--ignore-scripts` on post-release `npm install` (#2773)
As we are doing `--package-lock-only` (which does not actually install
packages), we don't have the dependencies needed to run the `install.js`
script, which is auto-run when installing. By passing
`--ignore-scripts`, we skip running `install.js`, and only end up
bumping the `package-lock.json`, which is what we want to do.