ci(rust): Run CI against toolchain update bot PRs (#2779)
The PRs from the Rust toolchain update bot don't trigger the
`pull_request`-triggered workflows. This is a [GitHub feature, not a
bug.](https://docs.github.com/en/actions/concepts/security/github_token?utm_source=chatgpt.com#when-github_token-triggers-workflow-runs)
This change adds a `workflow_dispatch` trigger to our CI action, and has
the Rust toolchain update bot automatically trigger the action after
opening the PR.

I tested this change with #2781. Although the workflow run does not show
on that PR, we can see that, in fact, the [CI did
run](https://github.com/getsentry/sentry-cli/actions/runs/17911030223)
against the commit. My guess is maybe it has to do with the fact that
the #2781 is against #2780, rather than the `master` branch.