feat(preprod): fail upload if app is missing Info.plist (#2793)
Every `.app` inside of the `Product/Applications` directory needs a
corresponding `Info.plist` file since we use this to gather important
information about the app. Currently we are allowing these uploads to
take place which then immediately fail during processing -- not the
greatest UX. It would be better to fail immediately so the user knows
there is an issue since this means something went wrong with packaging
their app.

Worth noting: technically there's a way to embed this file directly in
the Mach-O binary but that is seldom used and we don't support parsing
that in `launchpad` either.

Now we will fail and produce an error log:
```
  ERROR   2025-09-23 10:28:40.771928 -04:00 Invalid XCArchive: Missing required Info.plist file in .app bundle: /Users/telkins/Downloads/<redacted>.app.xcarchive/Products/Applications/<redacted>.app
error: File is not a recognized supported build format (APK, AAB, XCArchive, or IPA): /Users/telkins/Downloads/<redacted>.app.xcarchive
```

Resolves EME-311